requirements.txt
pandas
matplotlib
tensorflow
requests
pytz
grpcio
protobuf
types-protobuf
googleapis-common-protos
pyTelegramBotAPI
keras
numpy
Pillow
aiofiles
aiogram